\name{generic.plot}
\alias{generic.plot}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
generic.plot
}
\description{
Plots
}
\usage{
generic.plot(type = "R", width = 7, height = 7, name = "output", CALLFUN = NULL)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{type}{
"R", "png" ...
}
  \item{width}{
width of the file
}
  \item{height}{
height of the file
}
  \item{name}{
name of the file
}
  \item{CALLFUN}{
function to plot
}
}