\name{interieur.para}
\alias{interieur.para}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
interieur.para
}
\description{
Detects if a point is inside a given paralellogram
}
\usage{
interieur.para(x, y, xOrigine, yOrigine, xA, yA, xB = NA, yB = NA, shape = "para")
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
abcisse of the point to test
}
  \item{y}{
ordinate of the point to test
}
  \item{xOrigine}{
abscisse of a summit of the parallelogram  (origin)
}
  \item{yOrigine}{
ordinate of a summit of the parallelogram (origin)
}
  \item{xA}{
abscisse of a summit of the parallelogram linked to the origine
}
  \item{yA}{
ordinate of a summit of the parallelogram linked to the origine
}
  \item{xB}{
abscisse of the other summit of the parallelogram linked to the origine
}
  \item{yB}{
ordinate of the other summit of the parallelogram linked to the origine
}
  \item{shape}{
"para"
}
}